requirements.txt
telethon
python-telegram-bot==21.6
